refactor(openai-compatible): split query.ts into query/ submodules
openai-compatible/query.ts was the largest file in the repo (1403 LOC)
and never got the query/ decomposition anthropic-direct already has.
Extract the standalone helpers into cohesive modules; the session class
and turn loop stay in query.ts (now 1235 LOC).

- query/retry.ts: retryability predicates + computeBackoffDelay +
  __setRetryBaseDelay
- query/model-params.ts: output-cap, o-series, and effort helpers
- query/client.ts: resolveClientFactory + __setOpenAIClientFactory

query.ts re-exports the moved symbols so every existing import path
(sibling tests + index.ts) keeps resolving unchanged. Mutable
test-injection state is encapsulated behind computeBackoffDelay() and
resolveClientFactory() accessors. Behavior-preserving.
